+++ apps/playground/root/src/lib.rs
@@ -1,9 +1,0 @@
-#![expect(clippy::unused_async)]
-
-canic::start!();
-
-async fn canic_setup() {}
-async fn canic_install() {}
-async fn canic_upgrade() {}
-
-canic::finish!();